break up test-sanity into several files
i swear that was the intent from the beginning :-)
